+++ saleor/graphql/webhook/enums.py
@@ -3,6 +3,7 @@ import graphene
 from ...webhook.event_types import WebhookEventAsyncType, WebhookEventSyncType
 from ..core.descriptions import (
     DEFAULT_DEPRECATION_REASON,
+    DEPRECATED_LEGACY_PAYMENTS,
 )
 from ..core.doc_category import DOC_CATEGORY_WEBHOOKS
 from ..core.types import BaseEnum
@@ -285,6 +286,8 @@ def deprecation_reason(enum):
         )
     if enum.value == WebhookEventAsyncType.ANY:
         return DEFAULT_DEPRECATION_REASON
+    if enum.value in WebhookEventSyncType.PAYMENT_EVENTS:
+        return DEPRECATED_LEGACY_PAYMENTS
     return None